feat: 重构Shared.Host.Microservice
