+++ aspnet-core/gateways/Lion.AbpPro.WebGateway/GlobalUsings.cs
@@ -4,7 +4,6 @@ global using System;
 global using System.Collections.Generic;
 global using System.Linq;
 global using System.Threading.Tasks;
-global using Lion.AbpPro.Shared.Hosting.Gateways;
 global using Microsoft.AspNetCore.Builder;
 global using Microsoft.AspNetCore.Cors;
 global using Microsoft.AspNetCore.Hosting;
