+++ aspnet-core/services/host/Lion.AbpPro.HttpApi.Host/AbpProHttpApiHostModule.cs
@@ -1,3 +1,4 @@
+using Swagger;
 using Volo.Abp.BackgroundJobs.Hangfire;
 
 namespace Lion.AbpPro
@@ -18,8 +19,7 @@ namespace Lion.AbpPro
     )]
     public class AbpProHttpApiHostModule : AbpModule
     {
-        public override void OnPostApplicationInitialization(
-            ApplicationInitializationContext context)
+        public override void OnPostApplicationInitialization(ApplicationInitializationContext context)
         {
             // 应用程序初始化的时候注册hangfire
             context.CreateRecurringJob();
@@ -43,11 +43,14 @@ namespace Lion.AbpPro
         {
             var app = context.GetApplicationBuilder();
             var configuration = context.GetConfiguration();
-            //app.UseRequestLog();
             app.UseAbpRequestLocalization();
             app.UseCorrelationId();
             app.UseStaticFiles();
-            app.UseMiniProfiler();
+            if (configuration.GetValue("MiniProfiler:Enabled", false))
+            {
+                app.UseMiniProfiler();
+            }
+           
             app.UseRouting();
             app.UseCors(AbpProHttpApiHostConst.DefaultCorsPolicyName);
             app.UseAuthentication();
@@ -84,6 +87,8 @@ namespace Lion.AbpPro
         }
 
 
+        #region 私有配置
+
         private void ConfigureHangfire(ServiceConfigurationContext context)
         {
             Configure<AbpBackgroundJobOptions>(options => { options.IsJobExecutionEnabled = true; });
@@ -94,7 +99,7 @@ namespace Lion.AbpPro
                 var delaysInSeconds = new[] { 10, 60, 60 * 3 }; // 重试时间间隔
                 const int Attempts = 3; // 重试次数
                 config.UseFilter(new AutomaticRetryAttribute() { Attempts = Attempts, DelaysInSeconds = delaysInSeconds });
-                config.UseFilter(new AutoDeleteAfterSuccessAttributer(TimeSpan.FromDays(7)));
+                config.UseFilter(new AutoDeleteAfterSuccessAttribute(TimeSpan.FromDays(7)));
                 config.UseFilter(new JobRetryLastFilter(Attempts));
             });
         }
@@ -104,7 +109,11 @@ namespace Lion.AbpPro
         /// </summary>
         private void ConfigureMiniProfiler(ServiceConfigurationContext context)
         {
-            context.Services.AddMiniProfiler(options => options.RouteBasePath = "/profiler").AddEntityFramework();
+            if (context.Services.GetConfiguration().GetValue("MiniProfiler:Enabled", false))
+            {
+                context.Services.AddMiniProfiler(options => options.RouteBasePath = "/profiler").AddEntityFramework();
+            }
+          
         }
 
         /// <summary>
@@ -199,7 +208,7 @@ namespace Lion.AbpPro
             context.Services.Configure<IdentityOptions>(options => { options.Lockout = new LockoutOptions() { AllowedForNewUsers = false }; });
         }
 
-        private static void ConfigureSwaggerServices(ServiceConfigurationContext context)
+        private void ConfigureSwaggerServices(ServiceConfigurationContext context)
         {
             context.Services.AddSwaggerGen(
                 options =>
@@ -324,5 +333,7 @@ namespace Lion.AbpPro
                     options.IgnoredUrls.Add("/cap");
                 });
         }
+        
+        #endregion
     }
 }
