+++ aspnet-core/services/host/Lion.AbpPro.HttpApi.Host/Extensions/Hangfire/AutoDeleteAfterSuccessAttribute.cs
@@ -1,10 +1,10 @@
 ﻿namespace Lion.AbpPro.Extensions.Hangfire;
 
-public class AutoDeleteAfterSuccessAttributer : JobFilterAttribute, IApplyStateFilter
+public class AutoDeleteAfterSuccessAttribute : JobFilterAttribute, IApplyStateFilter
 {
     private readonly TimeSpan _deleteAfter;
 
-    public AutoDeleteAfterSuccessAttributer(TimeSpan timeSpan)
+    public AutoDeleteAfterSuccessAttribute(TimeSpan timeSpan)
     {
         _deleteAfter = timeSpan;
     }
