+++ aspnet-core/services/host/Lion.AbpPro.HttpApi.Host/GlobalUsings.cs
@@ -17,11 +17,7 @@ global using Lion.AbpPro.EntityFrameworkCore;
 global using Lion.AbpPro.Core;
 global using Lion.AbpPro.Extensions;
 global using Lion.AbpPro.Extensions.Hangfire;
-global using Lion.AbpPro.Extensions.Middlewares;
 global using Lion.AbpPro.MultiTenancy;
-global using Lion.AbpPro.Shared.Hosting.Microservices;
-global using Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.AspNetCore.Builder;
-global using Lion.AbpPro.Shared.Hosting.Microservices.Swaggers;
 global using Magicodes.ExporterAndImporter.Core;
 global using Magicodes.ExporterAndImporter.Excel;
 global using Microsoft.AspNetCore.Authentication.JwtBearer;
@@ -58,6 +54,7 @@ global using Volo.Abp.AspNetCore.Serilog;
 global using Volo.Abp.Auditing;
 global using Volo.Abp.Authorization;
 global using Volo.Abp.BackgroundJobs;
+global using Volo.Abp.BackgroundJobs.Hangfire;
 global using Volo.Abp.Caching;
 global using Volo.Abp.Caching.StackExchangeRedis;
 global using Volo.Abp.DependencyInjection;
