+++ aspnet-core/services/host/Lion.AbpPro.HttpApi.Host/appsettings.json
@@ -35,7 +35,7 @@
   },
   "ConnectionStrings": {
     "Default": "Data Source=localhost;Port=3306;Database=LionAbpProDB;uid=root;pwd=1q2w3E*;charset=utf8mb4;Allow User Variables=true;AllowLoadLocalInfile=true",
-    "Hangfire": "localhost:6379,defaultdatabase=1"
+    "Hangfire": "localhost:6379,defaultdatabase=5"
   },
   "Redis": {
     "Configuration": "localhost,password=1q2w3E*,defaultdatabase=1"
@@ -66,5 +66,8 @@
     "Enabled": false,
     "Host": "http://localhost:8500",
     "Service": "LionAbpPro-Service"
+  },
+  "MiniProfiler": {
+    "Enabled": true
   }
 }
