+++ aspnet-core/shared/Lion.AbpPro.Shared.Hosting.Gateways/Lion.AbpPro.Shared.Hosting.Gateways.csproj
@@ -2,6 +2,7 @@
     
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
+        <RootNamespace />
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Ocelot" />
