+++ aspnet-core/shared/Lion.AbpPro.Shared.Hosting.Microservices/GlobalUsings.cs
@@ -5,9 +5,10 @@ global using System.Collections.Generic;
 global using System.Linq;
 global using System.Net;
 global using System.Reflection;
+global using System.Text;
+global using System.Text.RegularExpressions;
 global using Consul;
 global using Lion.AbpPro.Core;
-global using Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.Extensions.DependencyInjection;
 global using Microsoft.AspNetCore.Builder;
 global using Microsoft.AspNetCore.Cors;
 global using Microsoft.AspNetCore.Http;
@@ -24,6 +25,9 @@ global using Microsoft.Extensions.Options;
 global using Microsoft.OpenApi.Any;
 global using Microsoft.OpenApi.Models;
 global using Newtonsoft.Json;
+global using Serilog;
+global using Serilog.Exceptions;
+global using Serilog.Sinks.Elasticsearch;
 global using Swashbuckle.AspNetCore.SwaggerGen;
 global using Volo.Abp;
 global using Volo.Abp.AspNetCore.ExceptionHandling;
