+++ aspnet-core/shared/Lion.AbpPro.Shared.Hosting.Microservices/Microsoft/AspNetCore/Builder/ApplicationBuilderExtensions.cs
@@ -1,8 +1,9 @@
-namespace Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.AspNetCore.Builder
+using Microsoft.AspNetCore.RequestLog;
+
+namespace Microsoft.AspNetCore.Builder
 {
     public static class ApplicationBuilderExtensions
     {
-        
         public static string UseConsul(this IApplicationBuilder app)
         {
             var appLifetime = app.ApplicationServices.GetService<IHostApplicationLifetime>();
@@ -21,7 +22,7 @@ namespace Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.AspNetCore.Builder
             string consulServiceId = $"{serviceName}:{serviceId}";
 
             var client = scope.ServiceProvider.GetService<IConsulClient>();
-            
+
             var consulServiceRegistration = new AgentServiceRegistration
             {
                 Name = serviceName,
@@ -30,10 +31,10 @@ namespace Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.AspNetCore.Builder
                 Port = appUrl.Port,
                 Check = new AgentServiceCheck
                 {
-                    DeregisterCriticalServiceAfter = TimeSpan.FromSeconds(5),//服务停止多久后注销
-                    Interval = TimeSpan.FromSeconds(3),//健康检查时间间隔，或者称为心跳 间隔
-                    HTTP = $"http://{appUrl.Host}:{appUrl.Port}/health",//健康检查地址 
-                    Timeout = TimeSpan.FromSeconds(15)   //超时时间
+                    DeregisterCriticalServiceAfter = TimeSpan.FromSeconds(5), //服务停止多久后注销
+                    Interval = TimeSpan.FromSeconds(3), //健康检查时间间隔，或者称为心跳 间隔
+                    HTTP = $"http://{appUrl.Host}:{appUrl.Port}/health", //健康检查地址 
+                    Timeout = TimeSpan.FromSeconds(15) //超时时间
                 }
             };
 
@@ -41,6 +42,15 @@ namespace Lion.AbpPro.Shared.Hosting.Microservices.Microsoft.AspNetCore.Builder
             appLifetime.ApplicationStopping.Register(() => { client.Agent.ServiceDeregister(consulServiceRegistration.ID); });
 
             return consulServiceId;
+        }
+
+        /// <summary>
+        /// 记录请求响应日志
+        /// </summary>
+        /// <returns></returns>
+        public static IApplicationBuilder UseRequestLog(this IApplicationBuilder app)
+        {
+            return app.UseMiddleware<RequestLogMiddleware>();
         }
     }
 }
