+++ aspnet-core/shared/Lion.AbpPro.Shared.Hosting.Microservices/Microsoft/AspNetCore/RequestLog/RequestLogMiddleware.cs
@@ -1,4 +1,4 @@
-﻿namespace Lion.AbpPro.Extensions.Middlewares;
+﻿namespace Microsoft.AspNetCore.RequestLog;
 
 public class RequestLogMiddleware
 {
